rating part done

+++ app/src/main/java/com/example/bait2073mobileapplicationdevelopment/entities/User.kt
@@ -230,7 +230,7 @@ data class UserRating(
     @ColumnInfo(name = "rating")
     @Expose
     @SerializedName("rating")
-    val rating: Float?,
+    val rating: Double?,
 
     )
